Review comments

+++ include/parsing/expression_parsing/math_lexer.hpp
@@ -59,7 +59,7 @@ struct MathLexerElement {
     std::string data;       ///< Data of the element
     int position;           ///< Position of the element in the input string
     int num_args;           ///< Number of args between current brackets (used for function argument separation)
-    int num_expressions;    ///< Number of expressions between current brackets (used for control flow constructs)
+    ptrdiff_t num_expressions;    ///< Number of expressions between current brackets (used for control flow constructs)
     /**
      * @brief Constructor for MathLexerElement.
      * @param type The type of the element.
@@ -68,14 +68,14 @@ struct MathLexerElement {
      * @param num_args The number of args between current brackets (used for function argument separation).
      * @param num_expressions The number of expressions between current brackets (used for control flow
      */
-    MathLexerElement(expression_type type, std::string data, int position, int num_args = -1, int num_expressions = -1)
+    MathLexerElement(expression_type type, std::string data, int position, int num_args = -1, ptrdiff_t num_expressions = -1)
         : type(type), data(data), position(position), num_args(num_args), num_expressions(num_expressions) {}
 
     void set_num_args(int num) {
         num_args = num;
     }
 
-    void set_num_expressions(int num) {
+    void set_num_expressions(ptrdiff_t num) {
         num_expressions = num;
     }
 };
